BookShop/requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2.1
django-rest-framework==0.1.0
django-taggit==1.1.0
djangorestframework==3.9.4
kombu==4.5.0
Markdown==3.1
Pillow==6.0.0
pkg-resources==0.0.0
PyMySQL==0.9.3
pytz==2019.1
sqlparse==0.3.0
vine==1.3.0
